requirements.txt
streamlit>=1.4.0
Pillow==9.0.0
pathlib==1.0.1